cool

+++ src/utility/point.py
@@ -1,8 +1,6 @@
 from dataclasses import dataclass
-from typing import Literal
-
 
 @dataclass
 class Point:
-    x: float
-    y: float
+    x: int
+    y: int
